# Conflicts:
#	integ-test/src/test/resources/expectedOutput/calcite/explain_agg_sort_on_metrics2.json
#	integ-test/src/test/resources/expectedOutput/calcite/explain_limit_agg_pushdown_bucket_nullable1.json
#	integ-test/src/test/resources/expectedOutput/calcite/explain_limit_agg_pushdown_bucket_nullable2.json

+++ integ-test/src/test/java/org/opensearch/sql/ppl/PrometheusDataSourceCommandsIT.java
@@ -207,7 +207,7 @@ public class PrometheusDataSourceCommandsIT extends PPLIntegTestCase {
                 + " span(@timestamp, 15s), handler, job");
     verifySchema(
         response,
-        schema("count()", "int"),
+        schema("count()", "bigint"),
         schema("span(@timestamp,15s)", "timestamp"),
         schema("handler", "string"),
         schema("job", "string"));
